fastapi/transcribe/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
openai==1.88.0
python-dotenv==1.0.1
pydub==0.25.1